change migration tools to TS because why not?
